Non-dyadic filter transforms
I've made a first pass at applying filter-based transforms to
non-dyadic signals. The level of transform is tested against the
length of the support. As long as the support has 2^level as a
factor, then the dwt can proceed.

Currently only implemented for dwt for 1D and 2D signals, using
filtering.

Still needs some testing.